Refactor Identity op to be part of qubit and cv modules (#1956)
* Refactor Identity op to be part of qubit and cv

* Changelog and format

* Add Identity to device tests

* Update doc/releases/changelog-dev.md


* Rename `common` file to `identity`

* Fix pylint unused import
