+++ pennylane/devices/tests/test_gates.py
@@ -37,6 +37,7 @@ np.random.seed(42)
 
 # gates for which device support is tested
 ops = {
+    "Identity": qml.Identity(wires=[0]),
     "BasisState": qml.BasisState(np.array([0]), wires=[0]),
     "CNOT": qml.CNOT(wires=[0, 1]),
     "CRX": qml.CRX(0, wires=[0, 1]),
