+++ pennylane/ops/__init__.py
@@ -18,10 +18,6 @@
 This module contains core quantum operations supported by PennyLane -
 such as gates, state preparations and observables.
 """
-import numpy as np
-
-from pennylane.operation import AnyWires, Observable, CVObservable, Operation
-
 from .cv import *
 from .qubit import *
 from .channel import *
@@ -43,52 +39,6 @@ class AdjointError(Exception):
     pass
 
 
-class Identity(CVObservable, Operation):
-    r"""pennylane.Identity(wires)
-    The identity observable :math:`\I`.
-
-    The expectation of this observable
-
-    .. math::
-        E[\I] = \text{Tr}(\I \rho)
-
-    corresponds to the trace of the quantum state, which in exact
-    simulators should always be equal to 1.
-    """
-    num_wires = 1
-    grad_method = None
-
-    ev_order = 1
-    eigvals = np.array([1, 1])
-
-    @property
-    def num_params(self):
-        return 0
-
-    def label(self, decimals=None, base_label=None):
-        return base_label or "I"
-
-    @classmethod
-    def _eigvals(cls, *params):
-        return cls.eigvals
-
-    @classmethod
-    def _matrix(cls, *params):
-        return np.eye(2)
-
-    @staticmethod
-    def _heisenberg_rep(p):
-        return np.array([1, 0, 0])
-
-    def diagonalizing_gates(self):
-        return []
-
-    @staticmethod
-    def identity_op(*params):
-        """Returns the matrix representation of the identity operator."""
-        return Identity._matrix(*params)
-
-
-__all__ = _cv__all__ + _qubit__all__ + _channel__ops__ + ["Identity"]
-__all_ops__ = list(_cv__ops__ | _qubit__ops__) + ["Identity"]
-__all_obs__ = list(_cv__obs__ | _qubit__obs__) + ["Identity"]
+__all__ = _cv__all__ + _qubit__all__ + _channel__ops__
+__all_ops__ = list(_cv__ops__ | _qubit__ops__)
+__all_obs__ = list(_cv__obs__ | _qubit__obs__)
